Move erros and watchdog code to separate file
