+++ crates/core/tedge_watchdog/src/error.rs
@@ -1,0 +1,30 @@
+use mqtt_channel::MqttError;
+
+use tedge_config::{ConfigSettingError, TEdgeConfigError};
+
+#[derive(Debug, thiserror::Error)]
+pub enum WatchdogError {
+    #[error("Fail to run `{cmd}`: {from}")]
+    CommandExecError { cmd: String, from: std::io::Error },
+
+    #[error(transparent)]
+    FromTedgeConfigError(#[from] TEdgeConfigError),
+
+    #[error(transparent)]
+    FromConfigSettingError(#[from] ConfigSettingError),
+
+    #[error(transparent)]
+    FromMqttError(#[from] MqttError),
+
+    #[error(transparent)]
+    DeserializeError(#[from] serde_json::Error),
+
+    #[error("Failed to parse watchdogsec")]
+    ParseWatchdogSec(#[from] std::num::ParseIntError),
+
+    #[error(transparent)]
+    ParseSystemdFile(#[from] std::io::Error),
+
+    #[error("Did not find the WatchdogSec")]
+    NoWatchdogSec,
+}
